requirements.txt
alembic==1.13.2
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
greenlet==3.0.3
h11==0.14.0
idna==3.7
Mako==1.3.5
MarkupSafe==2.1.5
outcome==1.3.0.post0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.23.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.32
SQLAlchemy-Utils==0.41.2
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
